Add InitFilesGenerator class for generating initialization files
